Stage

+++ src/TD.Komercijalno/TD.Komercijalno.Contracts/Dtos/Procedure/ProdajnaCenaNaDanDto.cs
@@ -2,7 +2,7 @@
 {
     public class ProdajnaCenaNaDanDto
     {
-        public int RobaId { get; set; }
+        public long RobaId { get; set; }
         public double ProdajnaCenaBezPDV { get; set; }
     }
 }
